refactor: centralize extractJson helper
