+++ backend/src/services/news-analyst.ts
@@ -1,5 +1,5 @@
 import { getNewsByToken } from '../repos/news.js';
-import { callAi } from '../util/ai.js';
+import { callAi, extractJson } from '../util/ai.js';
 import { analysisSchema, type Analysis } from './types.js';
 
 interface CacheEntry {
@@ -15,18 +15,6 @@ function getCacheKey(keyType: string, token: string) {
   return `${keyType}:${token}`;
 }
 
-function extractJson(res: string): Analysis | null {
-  try {
-    const json = JSON.parse(res);
-    const outputs = Array.isArray((json as any).output) ? (json as any).output : [];
-    const msg = outputs.find((o: any) => o.type === 'message' || o.id?.startsWith('msg_'));
-    const text = msg?.content?.[0]?.text;
-    if (typeof text !== 'string') return null;
-    return JSON.parse(text) as Analysis;
-  } catch {
-    return null;
-  }
-}
 
 export async function getTokenNewsSummary(
   token: string,
@@ -63,7 +51,7 @@ export async function getTokenNewsSummary(
       },
     };
     const res = await callAi(body, apiKey);
-    return extractJson(res);
+    return extractJson<Analysis>(res);
   })();
   cache.set(cacheKey, { promise, expires: now + THREE_MINUTES });
   try {
