+++ backend/src/services/order-book-analyst.ts
@@ -1,20 +1,7 @@
 import { fetchOrderBook } from './derivatives.js';
-import { callAi } from '../util/ai.js';
+import { callAi, extractJson } from '../util/ai.js';
 import { analysisSchema, type Analysis } from './types.js';
 import { setCache, getCache, acquireLock, releaseLock } from '../util/cache.js';
-
-function extractJson(res: string): Analysis | null {
-  try {
-    const json = JSON.parse(res);
-    const outputs = Array.isArray((json as any).output) ? (json as any).output : [];
-    const msg = outputs.find((o: any) => o.type === 'message' || o.id?.startsWith('msg_'));
-    const text = msg?.content?.[0]?.text;
-    if (typeof text !== 'string') return null;
-    return JSON.parse(text) as Analysis;
-  } catch {
-    return null;
-  }
-}
 
 export async function getOrderBookAnalysis(
   pair: string,
@@ -44,7 +31,7 @@ export async function getOrderBookAnalysis(
       },
     };
     const res = await callAi(body, apiKey);
-    const analysis = extractJson(res);
+    const analysis = extractJson<Analysis>(res);
     if (analysis) await setCache(key, analysis, 60 * 1000);
     return analysis;
   } finally {
